pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "policypulse-crawler"
version = "0.1.0"
description = "PolicyPulse news crawler and validator"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "PolicyPulse" }
]
dependencies = [
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.3",
  "lxml>=5.2.2",
  "feedparser>=6.0.11",
  "pyyaml>=6.0.1",
  "python-dateutil>=2.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["crawler"]
